+++ .idea/cssdialects.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="CssDialectMappings">
-    <file url="file://$PROJECT_DIR$/src/jisa/gui/primer-light.css" dialect="JavaFX" />
+    <file url="file://$PROJECT_DIR$/src/jisa/gui/style/breeze.css" dialect="JavaFX" />
+    <file url="file://$PROJECT_DIR$/src/jisa/gui/style/primer-light.css" dialect="JavaFX" />
   </component>
 </project>
